requirements.txt
# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
moto>=4.2.13  # For mocking AWS services in tests

# Core dependencies
pydantic>=2.9.2
python-dotenv==1.0.0  # For environment variable management
boto3>=1.35.74  # Required for Amazon Bedrock
PyMuPDF==1.22.5  # For PDF processing
pinecone-client==2.2.4  # For vector database
transformers==4.33.3  # For text processing
torch==2.0.1  # For deep learning
numpy>=1.26.4,<2.0.0  # Resolves conflict with langchain
pandas==2.0.3  # For data manipulation
scikit-learn==1.3.0  # For machine learning utilities
networkx==3.1  # For graph operations
tqdm==4.65.0  # For progress bars
python-magic==0.4.27  # For file type detection
python-magic-bin==0.4.14  # Windows binary for python-magic


# LangChain and related
langchain==0.3.18
langchain-openai==0.3.14
langchain-core==0.3.56
langchain-community==0.3.17
langchain-google-genai==2.0.8
langchain-aws==0.2.15
langchain-ollama==0.2.3
langchain_xai==0.2.0
langchain_experimental==0.3.4

# LLM Providers
openai==1.76.2
anthropic==0.8.0
ollama>=0.4.4,<1
whisper

# Database
neo4j>=5.17.0  # Neo4j Python driver
dynamodb-json==1.3.0  # For handling DynamoDB JSON formats

# Document processing
pypdf>=4.0.0  # For PDF file handling
pdf2image>=1.17.0  # For converting PDF to images
frontend # For document processing

# Attachment handling
imageio[ffmpeg]==2.37.0
pillow==11.1.0
ffmpeg==1.4.0
imagehash
docx2txt
pandas
fitz


# Additional dependencies
requests==2.32.2  # Required for API calls

# Captioning + Embedding + Vector DB
# (Already included: transformers, tqdm, pymupdf)

# NLP and Visualization
nltk>=3.9.1  # Natural Language Toolkit
matplotlib>=3.10.0  # For visualization

# Whisper
openai-whisper

# Rouge Score
rouge_score==0.1.2